separate filter class, add some unit tests
